Add job monitor to capture the delayed jobs/tasks
Add job monitor to capture the delayed jobs/tasks. If a task or a job takes more than 5 mins, then it is a Warning. If it is more than 10 mins, then it is an error. We can see the results on the screen or in a csv file. There is possibility to take the information for a task or a job by PID.